TimeSeries: add dtype param, default to float64 throughout
Promote the spline-storage layer to float64 end-to-end. The previous mix
of float64 sample arrays with float32 power-basis coeffs lost ~2-3 decimal
digits at fermentation timescales and forced an `_EPS = 1e-4` workaround
to absorb the quantization in event-spacing logic.

`TimeSeries` now has an explicit `dtype` field (default `jnp.float64`,
canonicalized via `jnp.dtype(...)`) governing every floating array field
(`times`, `values`, `jump_times`, `breaks`, `coeffs`). Mismatched-dtype
floating inputs to `__init__` warn-and-cast; binary ops require equal
dtypes and raise `TypeError` on mismatch. All methods that produce a new
`TimeSeries` (`deriv`, `__add__`/`__sub__`/`__mul__`/`__truediv__`,
`_binary_*`, `_fit_cubic_power_basis`) thread `self.dtype` through.

Storage helpers (`spline_ops`, `grid_utils`, `_convert_bspline_segments`,
`_fit_spline_timeseries`) compute at float64 throughout; the boundary
cast back to the user's requested dtype happens in the `TimeSeries`
constructor.

Serialization persists `"dtype"` as a numpy-canonical string from both
write paths (`timeseries_to_dict`, `_timeseries_to_dict_payload`) and
both read paths accept it; `from_dict` also takes a `dtype` kwarg that
overrides any saved field. Old payloads without `"dtype"` load as
float64.

The `_EPS` comment is rewritten to reflect physical event-spacing intent;
the value is unchanged.

Drive-by: drop a few unused imports (`VolumeChange`, `pseudobatch`) and
unused locals (`sv_at_meas`, `meas_t_np`) that ruff flagged while
editing.

Tests:
- New `tests/test_timeseries_dtype.py` (14 tests) covering default
  float64, explicit float32, casting warning, JSON roundtrip in both
  dtypes, backwards-compat (no `"dtype"` key), kwarg-overrides-saved,
  binary-op dtype-mismatch raises, and `evaluate_many(ts.times) ≈
  ts.values` at `rtol=1e-12` for cubic float64 splines.
- New `tests/conftest.py` enables `jax_enable_x64=True` for the suite so
  the float64 dtype is actually exercised (without it JAX silently
  downgrades).

Known issue surfaced by enabling x64:
`test_pseudospace_matches_segmented_with_sampling_and_bolus` now fails
with a 3e-3 volume divergence — investigated and found to be a
pre-existing latent bug, not a regression. Root cause is a float64
`linspace` rounding artefact (`linspace(0, 20, 181)[126] = 14.0+2ulp`)
interacting with the two integrators' different time-array conventions
and the test's nearest-neighbour resampler; `integrate_process_pseudospace`
puts that point post-event while `integrate_process` ends up pre-event,
and the round 0.003 L mismatch is exactly one sampling event's
bookkeeping. Float32 noise previously masked this. Documented in
`tmp/2026-04-25-LARGE-TASK-timeseries-dtype/NEXT-STEPS.md` with a
recommended fix (snap `t_eval` to event times in
`integrate_process_pseudospace`); not bundled here since it's
orthogonal to the dtype work.

+++ src/hybrax/format/time_series/grid_utils.py
@@ -18,11 +18,11 @@ def merge_times_with_tolerance(
     rtol: float = TIME_DEDUP_RTOL,
 ) -> jnp.ndarray:
     """Merge two sorted time arrays with tolerance-based deduplication."""
-    a = np.asarray(times_a, dtype=np.float32)
-    b = np.asarray(times_b, dtype=np.float32)
+    a = np.asarray(times_a, dtype=np.float64)
+    b = np.asarray(times_b, dtype=np.float64)
     candidate = np.sort(np.concatenate([a, b], axis=0))
     if candidate.size == 0:
-        return jnp.asarray([], dtype=jnp.float32)
+        return jnp.asarray([], dtype=jnp.float64)
 
     scale = float(candidate[-1] - candidate[0]) if candidate.size > 1 else 1.0
     if scale <= 0.0:
@@ -33,7 +33,7 @@ def merge_times_with_tolerance(
     for t in candidate[1:]:
         if float(t - merged[-1]) > tol:
             merged.append(float(t))
-    return jnp.asarray(np.asarray(merged, dtype=np.float32), dtype=jnp.float32)
+    return jnp.asarray(np.asarray(merged, dtype=np.float64), dtype=jnp.float64)
 
 
 def linear_interpolate_samples(
@@ -42,10 +42,10 @@ def linear_interpolate_samples(
     target_times: jnp.ndarray,
 ) -> jnp.ndarray:
     """Linearly interpolate source samples onto a target time grid."""
-    x = np.asarray(source_times, dtype=np.float32)
-    y = np.asarray(source_values, dtype=np.float32)
-    t = np.asarray(target_times, dtype=np.float32)
-    return jnp.asarray(np.interp(t, x, y).astype(np.float32), dtype=jnp.float32)
+    x = np.asarray(source_times, dtype=np.float64)
+    y = np.asarray(source_values, dtype=np.float64)
+    t = np.asarray(target_times, dtype=np.float64)
+    return jnp.asarray(np.interp(t, x, y).astype(np.float64), dtype=jnp.float64)
 
 
 def synthesize_binary_samples(
